Fixed issue with too long texts
Turns out, the problem was that when long messages didn't follow format, they were too long for Jocasta to DM in full back to the user. To fix, I split Jocasta's DM in 2 (one with the explanation, one with the original message), and added a character limit on the resent message that's just a few characters less to account for the backticks. Tested on my own bot, it works now